Rework chest to InteractableProp and  edit Level4
