+++ IoTether/Assets/Prefabs/Props/Chest.prefab
@@ -9,8 +9,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 5350398394705223075}
+  - component: {fileID: 1664351051302640914}
   - component: {fileID: 4172346870321244637}
-  - component: {fileID: 1462345086893824180}
   m_Layer: 0
   m_Name: Chest
   m_TagString: Untagged
@@ -36,6 +36,19 @@ Transform:
   - {fileID: 1786218336412603824}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1664351051302640914
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 564840342592710481}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b6f15ebf4e2d6ee4ea3add9c08d7a7d9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  chestDrop: {fileID: 0}
 --- !u!61 &4172346870321244637
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -81,19 +94,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1.8849955, y: 1.6062908}
   m_EdgeRadius: 0
---- !u!114 &1462345086893824180
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 564840342592710481}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a881b1b7bd0c38248a547fd8a065e6fd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chestDrop: {fileID: 7693310014522390134, guid: 3f5152ebbfcba78418ec0f0e3b35e69d, type: 3}
 --- !u!1 &786658879119129380
 GameObject:
   m_ObjectHideFlags: 0
