+++ IoTether/Assets/Scripts/Interfaces/InteractableProp.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b6f15ebf4e2d6ee4ea3add9c08d7a7d9
+guid: 1471ef27434bf55479b38ca28d3a6327
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
